+++ libffi-rs/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+- Add support for getting the offsets of a struct via `low::get_struct_offsets` or `middle::Type::struct_offsets`.
+
 ## [5.1.1] - 2026-06-21
 
 No user-facing changes.